Add ability to pass Queue objects to scheduler
Currently the scheduler looks for the queue with
`Queue.queue_from_key()`, which can be inadequate in cases where we want
a Queue object with properties that are not saved in Redis (for example,
`async=False`). This change makes it possible to bind the scheduler to a
specific queue, instead of looking for it by name and creating the
object.